LIFO Stack Resizing API using princeton algorithms library.
Stack resizing algorithm,

Resize
Resizes array by firstly creating a temp object with make sure that it casted to Item[] then, move it to the new max sized array.

Push
Takes an item parameter to add to the stack. If the stack's elements are equal to the array's length, it doubles the size to optimally make sure that needed amount is reserved.

Pop
Returns an Item object from the stack, and checks if the stack is too big it resizes  it to a smaller amount.